Remove unwanted Root.dev.js and Root.prod.js files
