+++ src/containers/Root.js
@@ -1,5 +1,16 @@
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./Root.prod')
-} else {
-  module.exports = require('./Root.dev')
+import React from 'react'
+import PropTypes from 'prop-types'
+import { Provider } from 'react-redux'
+import App from './App'
+
+const Root = props => (
+  <Provider store={props.store}>
+    <App />
+  </Provider>
+)
+
+Root.propTypes = {
+  store: PropTypes.shape({}).isRequired
 }
+
+export default Root
